style : 카테고리바 깨지는 스타일 수정
